# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kifayat Pakistan NGO API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactInputInquiryType = typeof ContactInputInquiryType[keyof typeof ContactInputInquiryType];


export const ContactInputInquiryType = {
  general: 'general',
  volunteer: 'volunteer',
  donation: 'donation',
  resource: 'resource',
  other: 'other',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @minLength 1 */
  message: string;
  inquiryType?: ContactInputInquiryType;
}

export type ContactSubmissionStatus = typeof ContactSubmissionStatus[keyof typeof ContactSubmissionStatus];


export const ContactSubmissionStatus = {
  new: 'new',
  read: 'read',
  replied: 'replied',
} as const;

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  inquiryType: string;
  status: ContactSubmissionStatus;
  createdAt: string;
}

export type ContactStatusUpdateStatus = typeof ContactStatusUpdateStatus[keyof typeof ContactStatusUpdateStatus];


export const ContactStatusUpdateStatus = {
  new: 'new',
  read: 'read',
  replied: 'replied',
} as const;

export interface ContactStatusUpdate {
  status: ContactStatusUpdateStatus;
}

export interface ContactList {
  items: ContactSubmission[];
  total: number;
  page: number;
  limit: number;
}

export type CampaignCategory = typeof CampaignCategory[keyof typeof CampaignCategory];


export const CampaignCategory = {
  shopkeepers: 'shopkeepers',
  vendors: 'vendors',
  laborers: 'laborers',
  general: 'general',
} as const;

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  active: 'active',
  completed: 'completed',
  upcoming: 'upcoming',
} as const;

export interface Campaign {
  id: number;
  title: string;
  description: string;
  category: CampaignCategory;
  status: CampaignStatus;
  beneficiaryCount: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  startDate?: string | null;
}

export type ResourceCategory = typeof ResourceCategory[keyof typeof ResourceCategory];


export const ResourceCategory = {
  shopkeepers: 'shopkeepers',
  vendors: 'vendors',
  laborers: 'laborers',
  general: 'general',
} as const;

export type ResourceResourceType = typeof ResourceResourceType[keyof typeof ResourceResourceType];


export const ResourceResourceType = {
  guide: 'guide',
  scheme: 'scheme',
  contact: 'contact',
  legal: 'legal',
  financial: 'financial',
} as const;

export interface Resource {
  id: number;
  title: string;
  description: string;
  category: ResourceCategory;
  resourceType: ResourceResourceType;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface Stats {
  totalContacts: number;
  newContacts: number;
  activeCampaigns: number;
  totalBeneficiaries: number;
  resourceCount: number;
  contactsByType?: CountByLabel[];
}

export type ListContactsParams = {
page?: number;
limit?: number;
};

export type ListResourcesParams = {
category?: string;
};