# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kifayat Pakistan NGO API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: campaigns
    description: NGO campaigns and initiatives
  - name: resources
    description: Resources for small retailers
  - name: stats
    description: Summary statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "400":
          description: Invalid input

  /admin/contacts:
    get:
      operationId: listContacts
      tags: [contact]
      summary: List all contact submissions (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of contact submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"

  /admin/contacts/{id}:
    get:
      operationId: getContact
      tags: [contact]
      summary: Get a single contact submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "404":
          description: Not found
    patch:
      operationId: updateContactStatus
      tags: [contact]
      summary: Update status of a contact submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all active campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"

  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources for small retailers
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get summary stats for the admin dashboard
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1
        inquiryType:
          type: string
          enum: [general, volunteer, donation, resource, other]
          default: general

    ContactSubmission:
      type: object
      required: [id, name, email, message, inquiryType, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        message:
          type: string
        inquiryType:
          type: string
        status:
          type: string
          enum: [new, read, replied]
        createdAt:
          type: string
          format: date-time

    ContactStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, read, replied]

    ContactList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactSubmission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    Campaign:
      type: object
      required: [id, title, description, category, status, beneficiaryCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [shopkeepers, vendors, laborers, general]
        status:
          type: string
          enum: [active, completed, upcoming]
        beneficiaryCount:
          type: integer
        imageUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date

    Resource:
      type: object
      required: [id, title, description, category, resourceType]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [shopkeepers, vendors, laborers, general]
        resourceType:
          type: string
          enum: [guide, scheme, contact, legal, financial]
        url:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]

    Stats:
      type: object
      required: [totalContacts, newContacts, activeCampaigns, totalBeneficiaries, resourceCount]
      properties:
        totalContacts:
          type: integer
        newContacts:
          type: integer
        activeCampaigns:
          type: integer
        totalBeneficiaries:
          type: integer
        resourceCount:
          type: integer
        contactsByType:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"

    CountByLabel:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer