# .migration-backup/lib/api-zod/src/generated/types/campaign.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kifayat Pakistan NGO API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignCategory } from './campaignCategory';
import type { CampaignStatus } from './campaignStatus';

export interface Campaign {
  id: number;
  title: string;
  description: string;
  category: CampaignCategory;
  status: CampaignStatus;
  beneficiaryCount: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  startDate?: Date | null;
}