# .migration-backup/lib/api-zod/src/generated/types/contactStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kifayat Pakistan NGO API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactStatusUpdateStatus } from './contactStatusUpdateStatus';

export interface ContactStatusUpdate {
  status: ContactStatusUpdateStatus;
}