# .migration-backup/lib/api-zod/src/generated/types/countByLabel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kifayat Pakistan NGO API
 * OpenAPI spec version: 0.1.0
 */

export interface CountByLabel {
  label: string;
  count: number;
}