# .migration-backup/lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kifayat Pakistan NGO API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceCategory } from './resourceCategory';
import type { ResourceResourceType } from './resourceResourceType';

export interface Resource {
  id: number;
  title: string;
  description: string;
  category: ResourceCategory;
  resourceType: ResourceResourceType;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
}